move assert_result to its own file
